yn統一
